// src/common/types/IMovie.ts
    export  interface IMovie {
        Title: string;
        Year: string;
        Poster: string;
        ID: string;
        Type: string;
    }

    export  interface IMovieAggregate {
        Title: string;
        Year: string;
        Poster: string;
        FilmWorldID: string;
        CinemaWorldID:string;
        Type: string;
    }